requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
langchain-text-splitters
sentence_transformers
langchain_huggingface
streamlit
langchain_groq
langchain_core